Moved the action selector to it's own file
